fix: guard lazy perps-worth load and shrink bg startup graph further
- Wrap the perps-worth lazy load in try/catch at the call site and drop a
  failed module load from the cached promise so the next call retries;
  a chunk-load failure no longer breaks the batch's tokens/DeFi values
  or pins the rejection on the bg singleton for the whole session
- Move fixOthersWalletAccountNetworkPair and mergeHomeDataToSwapMap bodies
  into accountSelectorHomeSyncUtils and load that module only via dynamic
  import, taking it out of the native background startup graph
